requirements.txt
streamlit
pysqlite3-binary
chromadb

openai==1.10.0
langchain==0.1.3
langchain_openai==0.0.5
langchain_community==0.0.16
PyPDF2==3.0.1
bs4==0.0.1
requests==2.31.0
tiktoken==0.5.2



